move test-hooks
