+++ test/smoke/src/areas/positron/console/python-console.test.ts
@@ -5,7 +5,7 @@
 
 import { expect } from '@playwright/test';
 import { Application, PositronPythonFixtures } from '../../../../../automation';
-import { setupAndStartApp } from '../../../test-hooks';
+import { setupAndStartApp } from '../../../test-runner/test-hooks';
 
 describe('Console Pane: Python #web #win', () => {
 	setupAndStartApp();
